verbose EXPERT level
With this addition I tried to make the autocalibration more universal to
all delta's. Playing with the routine on my Flsun and by the feedback in
the Flsun FB group, I 'discovered' that diverting from the theoretical
convertion matrix (= linear approximation to the differential of the
equations) the autocalibration converges much faster without losing
precision in the least squares approach. This is logical with non-linear
equations. Non shure however if the ideal Flsun matrices will work on
all geometries. That is why I parametered them in the previous commit.
In this commit I'm making an attempt to approach this problem more
scientificly then try and error, by comparing the actual changes that
are made to the changes that should be made by the convergence matrices
(the error so to speak of the linear approximation to the non-linear
problem) and calculating the matrix parameters to be used to reduce this
'error'. This is a process that should be done once by an 'expert' for
each brand (kossel-mini, kossel-xl,...) and then set it and forget it.  So
I disabled the expert-routines by default; 'experts' should enable it in
configuration_adv.h; run the tests; override the default values with the
ones obtained; and disable the expert-routines after that. I only tested
it on the Flsun configuration, but it seems to work and give the same
results I obtained by months of try and error. This however needs to be
confirmed by more testing, especially on bigger deltas.