Rework the STM32F7xx HAL as a generic STM32 HAL
The offical arduino STM32 core is abstracting quite much for us
already, so the only real device specific parts are for the timer
configuration.

Memory backtrace information should be handled on a motherboard
basis, since STM32 chips are pretty configurable, as you can already
see in the difference between the chips used in THE_BORG and REMRAM_V1.

If a board wants to use this HAL, it needs a variant definition
in the upstream STM32 core.